Handle sinuoide data send a
