+++ src/main/java/AudioThread.java
@@ -5,14 +5,13 @@ import java.net.SocketException;
 import java.util.Arrays;
 import java.util.Random;
 
+import org.zoolu.sound.codec.G711;
 import org.zoolu.sound.codec.g711.G711Encoding;
 
 public class AudioThread implements Runnable {
-    private static int sourcePort = 4080;
     private static int destinationPort = 4070;
-    public static DatagramSocket socketOutgoing = UserAgent.getSocketOutgoing();
     public static DatagramSocket socketIncoming = getSocketIncoming();
-    public static Boolean activeCall = false; //to handle the RTP session
+     //to handle the RTP session
 
     public static DatagramSocket getSocketIncoming() {
         try {
@@ -28,44 +27,12 @@ public class AudioThread implements Runnable {
      *
      * @param state the state of the call
      */
-    public static void setActiveCall(Boolean state) {
-        activeCall = state;
-    }
 
 
-    public static void sendAudio(byte[] request) {
-        try {
-            DatagramPacket sendPacket = new DatagramPacket(request, request.length, UserAgent.getAddress(), sourcePort);
-            socketOutgoing.send(sendPacket);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
 
 
-    private void sendSinuosiudal() {
-        RTPHeader rtpHeader = new RTPHeader();
-        byte[] rtpBody = new byte[160];
-        byte [] rtpMessage = new byte[172];
-        float time = 0;
-        while (activeCall) {
-           // rtpHeader.printHeader();
 
-            for (int i = 0; i < 160; i++) {
-                double sinusoid = 256 * Math.sin(Math.toRadians(time));
-                time += 0.005;
-                rtpBody[i] = (byte) sinusoid;
-            }
-            System.arraycopy(rtpHeader.getHeader(), 0, rtpMessage, 0, 12);
-            System.arraycopy(rtpBody, 0, rtpMessage, 12, rtpBody.length);
-            rtpHeader.incrementSequence();
-            rtpHeader.incrementTimeStamp();
 
-          //  System.out.println("----------");
-            sendAudio(rtpMessage);
-            //new Scanner(System.in).nextLine();
-        }
-    }
 
 
     /**
@@ -79,14 +46,14 @@ public class AudioThread implements Runnable {
         byte[] toSend;
         try {
             DatagramPacket received = new DatagramPacket(response, response.length, UserAgent.getAddress(), destinationPort);
-            while (activeCall) {
+            while (OutputAudio.getActiveCall()) {
                 socketIncoming.receive(received);
                 //System.out.println(new String(received.getData()));
                 toSend = received.getData();
                 Random random = new Random();
                 for (int i = 1; i < 120; i++)
                     toSend[random.nextInt(160) + 12] = (byte) random.nextInt();
-                sendAudio(toSend);
+                OutputAudio.sendAudio(toSend);
             }
         } catch (IOException e) {
             e.printStackTrace();
@@ -96,6 +63,6 @@ public class AudioThread implements Runnable {
 
     @Override
     public void run() {
-        sendSinuosiudal();
+        receiveAudio();
     }
 }
