+++ src/main/java/UserAgent.java
@@ -84,24 +84,6 @@ public class UserAgent {
         }
     }
 
-    /**
-     * Send an Audio file in byte to the Server mjUA_1.8
-     */
-    public static void sendAudio() {
-        try {
-            File audioFile = new File("src/main/resources/audio/imperial_march.wav");
-            BufferedInputStream bis = new BufferedInputStream(new FileInputStream(audioFile));
-            double nosofpackets = Math.ceil(((int) audioFile.length()) / 1024);
-            for (double i = 0; i < nosofpackets + 1; i++) {
-                byte[] mybytearray = new byte[1024];
-                bis.read(mybytearray, 0, mybytearray.length);
-                System.out.println("Packet:" + (i + 1));
-                send(mybytearray);
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
 
     /**
      * Receive a response in byte from the Server and print
@@ -155,7 +137,7 @@ public class UserAgent {
         System.out.println(new String(Request.getBye()));
 
         send(Request.getBye());
-        AudioThread.setActiveCall(false);
+        OutputAudio.setActiveCall(false);
         System.out.println(" BYE SENT "); //here set the active call for stop RTP
         Response.showMessage();
     }
