+++ src/main/resources/requests/ack.txt
@@ -1,9 +1,9 @@
 ACK sip:bob@127.0.0.1:5080 SIP/2.0
-Via: SIP/2.0/UDP 127.0.0.1:5070;branch=z9hG4bK62378an3
+Via: SIP/2.0/UDP 127.0.0.1:5070;branch=z9hG4bKj658em7t
 Max-Forwards: 70
-To: "Bob" <sip:bob@127.0.0.1:5080>;tag=297f62463b8a9b85
-From: "Alice" <sip:alice@127.0.0.1:5070>;tag=568518076493
-Call-ID: 126990851630@127.0.0.1
+To: "Bob" <sip:bob@127.0.0.1:5080>;tag=7879f2b5f26d4a33
+From: "Alice" <sip:alice@127.0.0.1:5070>;tag=604937336972
+Call-ID: 509710066208@127.0.0.1
 CSeq: 1 ACK
 Contact: <sip:alice@127.0.0.1:5070>
 Expires: 3600
