+++ src/main/resources/requests/bye.txt
@@ -1,9 +1,9 @@
 BYE sip:bob@127.0.0.1:5080 SIP/2.0
-Via: SIP/2.0/UDP 127.0.0.1:5070;branch=z9hG4bKyarcvwql
+Via: SIP/2.0/UDP 127.0.0.1:5070;branch=z9hG4bKuf5g3cau
 Max-Forwards: 70
-To: "Bob" <sip:bob@127.0.0.1:5080>;tag=297f62463b8a9b85
-From: "Alice" <sip:alice@127.0.0.1:5070>;tag=568518076493
-Call-ID: 126990851630@127.0.0.1
+To: "Bob" <sip:bob@127.0.0.1:5080>;tag=6275b1aa356744d8
+From: "Alice" <sip:alice@127.0.0.1:5070>;tag=386788402615
+Call-ID: 682864731775@127.0.0.1
 CSeq: 2 BYE
 User-Agent: mjsip 1.8
 Content-Length: 0
